backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
pandas
openpyxl
PyMuPDF
pytesseract
Pillow
pdf2image
